RF Unit classes and moved Tests to ViewsTest
